# Improve FTS indexing performance