--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -30,7 +30,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v3
       - name: Install Postgres Server for Tests
-        run: sudo apt install -y postgresql
+        run: sudo apt install -y postgresql postgresql-17-pgvector
       - name: go tests
         run: go test -v -covermode=count -json ./... > test.json
       - name: annotate go tests