--- a/example/go.mod
+++ b/example/go.mod
@@ -1,24 +1,24 @@
 module github.com/ductone/protoc-gen-pgdb/example
 
-go 1.24
+go 1.24.0
 
 replace github.com/ductone/protoc-gen-pgdb => ../
 
 require (
 	github.com/doug-martin/goqu/v9 v9.19.0
-	github.com/ductone/protoc-gen-pgdb v0.4.9
-	github.com/jackc/pgx/v5 v5.7.4
+	github.com/ductone/protoc-gen-pgdb v0.6.1
+	github.com/jackc/pgx/v5 v5.7.5
 	github.com/pquerna/protoc-gen-dynamo v0.8.0
 	github.com/segmentio/ksuid v1.0.4
 	github.com/stretchr/testify v1.8.1
-	google.golang.org/protobuf v1.36.6
+	google.golang.org/protobuf v1.36.8
 )
 
 require (
 	github.com/clipperhouse/jargon v1.0.9 // indirect
-	github.com/clipperhouse/uax29 v1.14.3 // indirect
+	github.com/clipperhouse/uax29 v1.15.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/gaissmai/extnetip v1.1.0 // indirect
+	github.com/gaissmai/extnetip v1.2.2 // indirect
 	github.com/jackc/pgpassfile v1.0.0 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
 	github.com/jackc/puddle/v2 v2.2.2 // indirect
@@ -27,10 +27,10 @@ require (
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/crypto v0.38.0 // indirect
-	golang.org/x/exp v0.0.0-20250506013437-ce4c2cf36ca6 // indirect
-	golang.org/x/net v0.40.0 // indirect
-	golang.org/x/sync v0.14.0 // indirect
-	golang.org/x/text v0.25.0 // indirect
+	golang.org/x/crypto v0.41.0 // indirect
+	golang.org/x/exp v0.0.0-20250819193227-8b4c13bb791b // indirect
+	golang.org/x/net v0.43.0 // indirect
+	golang.org/x/sync v0.16.0 // indirect
+	golang.org/x/text v0.28.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )