- Improve FTS Edgegram code, much faster and lower allocations for large strings.
- Upgrade `jargon` to fork that supports updated deps / golang 1.25
```
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=1-8              3672          3034          -17.37%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=2-8              3321          2832          -14.72%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=3-8              3145          2711          -13.80%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=5-8              3145          2671          -15.07%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=8-8              3216          2675          -16.82%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=16-8             3139          2655          -15.42%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=1-8              3708          3181          -14.21%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=2-8              3360          2924          -12.98%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=3-8              3169          2701          -14.77%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=5-8              3155          2684          -14.93%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=8-8              3142          2669          -15.05%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=16-8             3135          2684          -14.39%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=1-8          5992          4150          -30.74%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=2-8          5663          3945          -30.34%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=3-8          5438          3763          -30.80%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=5-8          4498          3370          -25.08%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=8-8          4024          3219          -20.00%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=16-8         3996          3213          -19.59%
BenchmarkEdgegramStream_Pipeline/Sentence/n=1-8                    8394          6042          -28.02%
BenchmarkEdgegramStream_Pipeline/Sentence/n=2-8                    7373          5797          -21.38%
BenchmarkEdgegramStream_Pipeline/Sentence/n=3-8                    7077          5218          -26.27%
BenchmarkEdgegramStream_Pipeline/Sentence/n=5-8                    5884          4313          -26.70%
BenchmarkEdgegramStream_Pipeline/Sentence/n=8-8                    5858          4296          -26.66%
BenchmarkEdgegramStream_Pipeline/Sentence/n=16-8                   5865          4360          -25.66%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=1-8                  6288          4872          -22.52%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=2-8                  5345          4419          -17.32%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=3-8                  5344          4352          -18.56%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=5-8                  5334          4359          -18.28%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=8-8                  5362          4390          -18.13%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=16-8                 5338          4369          -18.15%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=1-8            4256          3728          -12.41%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=2-8            4031          3600          -10.69%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=3-8            3889          3413          -12.24%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=5-8            3479          3100          -10.89%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=8-8            3459          3117          -9.89%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=16-8           3458          3107          -10.15%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=1-8                    2949          2828          -4.10%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=2-8                    2903          2692          -7.27%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=3-8                    2900          2780          -4.14%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=5-8                    2903          2688          -7.41%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=8-8                    2912          2733          -6.15%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=16-8                   2897          2684          -7.35%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=1-8                7603          5842          -23.16%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=2-8                7444          5609          -24.65%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=3-8                7500          5795          -22.73%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=5-8                6997          5138          -26.57%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=8-8                6013          4505          -25.08%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=16-8               6012          4505          -25.07%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=1-8              5077657       420492        -91.72%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=2-8              5203545       416509        -92.00%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=3-8              5210387       417882        -91.98%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=5-8              5342315       414036        -92.25%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=8-8              5279759       407408        -92.28%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=16-8             5321394       414379        -92.21%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=1-8              45069633      2320977       -94.85%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=2-8              45131379      2242826       -95.03%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=3-8              45187012      2231039       -95.06%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=5-8              45503903      2293198       -94.96%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=8-8              45279383      2331251       -94.85%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=16-8             45670811      2330734       -94.90%
BenchmarkEdgegramStream_Pipeline/Paragraph_English_2KiB/n=1-8      945864        586723        -37.97%
BenchmarkEdgegramStream_Pipeline/Paragraph_English_2KiB/n=2-8      772029        518065        -32.90%
BenchmarkEdgegramStream_Pipeline/Paragraph_English_2KiB/n=3-8      692722        441428        -36.28%
BenchmarkEdgegramStream_Pipeline/Paragraph_English_2KiB/n=5-8      599962        372888        -37.85%
BenchmarkEdgegramStream_Pipeline/Paragraph_English_2KiB/n=8-8      600998        372527        -38.02%
BenchmarkEdgegramStream_Pipeline/Paragraph_English_2KiB/n=16-8     598516        370873        -38.03%
BenchmarkEdgegramStream_Isolated/SmallASCII_git/n=1-8              1841          1171          -36.39%
BenchmarkEdgegramStream_Isolated/SmallASCII_git/n=2-8              1523          1054          -30.79%
BenchmarkEdgegramStream_Isolated/SmallASCII_git/n=3-8              1351          916           -32.20%
BenchmarkEdgegramStream_Isolated/SmallASCII_git/n=5-8              1355          886           -34.59%
BenchmarkEdgegramStream_Isolated/SmallASCII_git/n=8-8              1363          894           -34.39%
BenchmarkEdgegramStream_Isolated/SmallASCII_git/n=16-8             1351          885           -34.46%
BenchmarkEdgegramStream_Isolated/SmallASCII_hub/n=1-8              1849          1253          -32.23%
BenchmarkEdgegramStream_Isolated/SmallASCII_hub/n=2-8              1575          1076          -31.68%
BenchmarkEdgegramStream_Isolated/SmallASCII_hub/n=3-8              1352          916           -32.23%
BenchmarkEdgegramStream_Isolated/SmallASCII_hub/n=5-8              1351          890           -34.09%
BenchmarkEdgegramStream_Isolated/SmallASCII_hub/n=8-8              1356          888           -34.48%
BenchmarkEdgegramStream_Isolated/SmallASCII_hub/n=16-8             1366          886           -35.16%
BenchmarkEdgegramStream_Isolated/MediumASCII_github/n=1-8          3789          2101          -44.55%
BenchmarkEdgegramStream_Isolated/MediumASCII_github/n=2-8          3532          1815          -48.61%
BenchmarkEdgegramStream_Isolated/MediumASCII_github/n=3-8          3346          1696          -49.31%
BenchmarkEdgegramStream_Isolated/MediumASCII_github/n=5-8          2505          1407          -43.83%
BenchmarkEdgegramStream_Isolated/MediumASCII_github/n=8-8          2098          1271          -39.42%
BenchmarkEdgegramStream_Isolated/MediumASCII_github/n=16-8         2066          1270          -38.53%
BenchmarkEdgegramStream_Isolated/Sentence/n=1-8                    6971          3999          -42.63%
BenchmarkEdgegramStream_Isolated/Sentence/n=2-8                    7517          4277          -43.10%
BenchmarkEdgegramStream_Isolated/Sentence/n=3-8                    7147          4001          -44.02%
BenchmarkEdgegramStream_Isolated/Sentence/n=5-8                    6867          3415          -50.27%
BenchmarkEdgegramStream_Isolated/Sentence/n=8-8                    5260          2862          -45.59%
BenchmarkEdgegramStream_Isolated/Sentence/n=16-8                   4151          1814          -56.30%
BenchmarkEdgegramStream_Isolated/Punctuated/n=1-8                  6073          3395          -44.10%
BenchmarkEdgegramStream_Isolated/Punctuated/n=2-8                  6428          3604          -43.93%
BenchmarkEdgegramStream_Isolated/Punctuated/n=3-8                  6640          3633          -45.29%
BenchmarkEdgegramStream_Isolated/Punctuated/n=5-8                  5881          2848          -51.57%
BenchmarkEdgegramStream_Isolated/Punctuated/n=8-8                  4441          2417          -45.58%
BenchmarkEdgegramStream_Isolated/Punctuated/n=16-8                 3905          1761          -54.90%
BenchmarkEdgegramStream_Isolated/UnicodeCombining/n=1-8            2314          1768          -23.60%
BenchmarkEdgegramStream_Isolated/UnicodeCombining/n=2-8            2090          1631          -21.96%
BenchmarkEdgegramStream_Isolated/UnicodeCombining/n=3-8            1949          1473          -24.42%
BenchmarkEdgegramStream_Isolated/UnicodeCombining/n=5-8            1616          1243          -23.08%
BenchmarkEdgegramStream_Isolated/UnicodeCombining/n=8-8            1619          1244          -23.16%
BenchmarkEdgegramStream_Isolated/UnicodeCombining/n=16-8           1621          1257          -22.46%
BenchmarkEdgegramStream_Isolated/EmojiZWJ/n=1-8                    1160          1083          -6.64%
BenchmarkEdgegramStream_Isolated/EmojiZWJ/n=2-8                    1164          958           -17.72%
BenchmarkEdgegramStream_Isolated/EmojiZWJ/n=3-8                    1165          958           -17.73%
BenchmarkEdgegramStream_Isolated/EmojiZWJ/n=5-8                    1162          963           -17.13%
BenchmarkEdgegramStream_Isolated/EmojiZWJ/n=8-8                    1160          959           -17.35%
BenchmarkEdgegramStream_Isolated/EmojiZWJ/n=16-8                   1160          966           -16.69%
BenchmarkEdgegramStream_Isolated/MixedScripts/n=1-8                5530          3668          -33.67%
BenchmarkEdgegramStream_Isolated/MixedScripts/n=2-8                6089          3932          -35.42%
BenchmarkEdgegramStream_Isolated/MixedScripts/n=3-8                6264          3775          -39.73%
BenchmarkEdgegramStream_Isolated/MixedScripts/n=5-8                5024          3417          -31.99%
BenchmarkEdgegramStream_Isolated/MixedScripts/n=8-8                4201          2639          -37.18%
BenchmarkEdgegramStream_Isolated/MixedScripts/n=16-8               3788          2214          -41.55%
BenchmarkEdgegramStream_Isolated/LongASCII_1KiB/n=1-8              5019966       357407        -92.88%
BenchmarkEdgegramStream_Isolated/LongASCII_1KiB/n=2-8              5126797       352941        -93.12%
BenchmarkEdgegramStream_Isolated/LongASCII_1KiB/n=3-8              5135115       351788        -93.15%
BenchmarkEdgegramStream_Isolated/LongASCII_1KiB/n=5-8              5226667       349340        -93.32%
BenchmarkEdgegramStream_Isolated/LongASCII_1KiB/n=8-8              5345489       341693        -93.61%
BenchmarkEdgegramStream_Isolated/LongASCII_1KiB/n=16-8             5242187       357300        -93.18%
BenchmarkEdgegramStream_Isolated/LongASCII_4KiB/n=1-8              44788335      2305566       -94.85%
BenchmarkEdgegramStream_Isolated/LongASCII_4KiB/n=2-8              44928791      2260683       -94.97%
BenchmarkEdgegramStream_Isolated/LongASCII_4KiB/n=3-8              44996491      2271715       -94.95%
BenchmarkEdgegramStream_Isolated/LongASCII_4KiB/n=5-8              45650439      2297083       -94.97%
BenchmarkEdgegramStream_Isolated/LongASCII_4KiB/n=8-8              45831271      2327993       -94.92%
BenchmarkEdgegramStream_Isolated/LongASCII_4KiB/n=16-8             45326508      2315771       -94.89%
BenchmarkEdgegramStream_Isolated/Paragraph_English_2KiB/n=1-8      17012776      1055946       -93.79%
BenchmarkEdgegramStream_Isolated/Paragraph_English_2KiB/n=2-8      16835714      1023121       -93.92%
BenchmarkEdgegramStream_Isolated/Paragraph_English_2KiB/n=3-8      17012737      1075621       -93.68%
BenchmarkEdgegramStream_Isolated/Paragraph_English_2KiB/n=5-8      17048673      1088412       -93.62%
BenchmarkEdgegramStream_Isolated/Paragraph_English_2KiB/n=8-8      17209051      1076049       -93.75%
BenchmarkEdgegramStream_Isolated/Paragraph_English_2KiB/n=16-8     17022143      1035376       -93.92%

benchmark                                                          old allocs     new allocs     delta
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=1-8              47             35             -25.53%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=2-8              42             33             -21.43%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=3-8              40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=5-8              40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=8-8              40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_git/n=16-8             40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=1-8              47             35             -25.53%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=2-8              43             34             -20.93%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=3-8              40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=5-8              40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=8-8              40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/SmallASCII_hub/n=16-8             40             32             -20.00%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=1-8          66             45             -31.82%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=2-8          67             47             -29.85%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=3-8          65             46             -29.23%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=5-8          55             41             -25.45%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=8-8          50             39             -22.00%
BenchmarkEdgegramStream_Pipeline/MediumASCII_github/n=16-8         50             39             -22.00%
BenchmarkEdgegramStream_Pipeline/Sentence/n=1-8                    100            62             -38.00%
BenchmarkEdgegramStream_Pipeline/Sentence/n=2-8                    96             66             -31.25%
BenchmarkEdgegramStream_Pipeline/Sentence/n=3-8                    90             62             -31.11%
BenchmarkEdgegramStream_Pipeline/Sentence/n=5-8                    78             54             -30.77%
BenchmarkEdgegramStream_Pipeline/Sentence/n=8-8                    78             54             -30.77%
BenchmarkEdgegramStream_Pipeline/Sentence/n=16-8                   78             54             -30.77%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=1-8                  87             57             -34.48%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=2-8                  74             53             -28.38%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=3-8                  74             53             -28.38%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=5-8                  74             53             -28.38%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=8-8                  74             53             -28.38%
BenchmarkEdgegramStream_Pipeline/Punctuated/n=16-8                 74             53             -28.38%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=1-8            52             40             -23.08%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=2-8            51             41             -19.61%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=3-8            48             39             -18.75%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=5-8            43             36             -16.28%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=8-8            43             36             -16.28%
BenchmarkEdgegramStream_Pipeline/UnicodeCombining/n=16-8           43             36             -16.28%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=1-8                    30             28             -6.67%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=2-8                    30             26             -13.33%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=3-8                    30             26             -13.33%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=5-8                    30             26             -13.33%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=8-8                    30             26             -13.33%
BenchmarkEdgegramStream_Pipeline/EmojiZWJ/n=16-8                   30             26             -13.33%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=1-8                92             64             -30.43%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=2-8                94             63             -32.98%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=3-8                94             65             -30.85%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=5-8                88             61             -30.68%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=8-8                79             57             -27.85%
BenchmarkEdgegramStream_Pipeline/MixedScripts/n=16-8               79             57             -27.85%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=1-8              4158           2082           -49.93%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=2-8              5179           2082           -59.80%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=3-8              5177           2082           -59.78%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=5-8              5173           2082           -59.75%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=8-8              5167           2082           -59.71%
BenchmarkEdgegramStream_Pipeline/LongASCII_1KiB/n=16-8             5151           2082           -59.58%
BenchmarkEdgegramStream_Pipeline/LongASCII_4KiB/n=1-8              12371          6187           -49.99%
…